requirements.txt
Flask==3.1.2
pymongo[srv]==4.16.0
python-dotenv==1.1.1
gunicorn==25.1.0
